LevelSO Resetter

+++ Assets/Scripts/Action System/FerocityBehave.cs
@@ -6,7 +6,7 @@ public class FerocityBehave : BehaveVariable
 {
     [Tooltip("The % of hp a foe must have left to trigger this added weight")]
     [SerializeField]
-    int hpThresholdPercent;
+    float hpThresholdPercent;
     public override void ConditionallyApplyMod()
     {
         if(affectedBehaviour.actionVariations.Count == 0)
@@ -20,7 +20,7 @@ public class FerocityBehave : BehaveVariable
             if (!tgt)
                 continue;
 
-            if (tgt.currentHP <= tgt.maxHP / (100/hpThresholdPercent))
+            if (tgt.currentHP <= tgt.maxHP / (100f/hpThresholdPercent))
             {
                 av.weight *= modifier;
                 Debug.Log("FEROCITY ADDED!");
